requirements/llava.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
bitsandbytes==0.41.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6
datasets==2.18.0
einops==0.6.1
exceptiongroup==1.2.2
numpy==1.26.1
openai==0.28.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.7
propcache==0.3.0
protobuf==3.20
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==1.10.8
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.15.2
sentencepiece==0.1.99
sentry-sdk==2.24.1
setproctitle==1.3.5
setuptools==75.8.0
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.50.1
triton==2.1.0
flash-attn==2.3.4
gdown
git+https://github.com/LLaVA-VL/LLaVA-NeXT.git